refactor(core): extract zoom facade members for max-lines gate
Move getZoom, setZoom, getZoomMode, setZoomMode, getRenderScale, and
getPageGeometry into docx-editor-zoom.ts so docx-editor.ts stays under
the 2600-line lint cap after merging main.
