+++ docs/api/docx-editor-core/output.api.md
@@ -5,12 +5,6 @@
 ```ts
 
 // @public
-export const DEFAULT_FIELD_SHADING: FieldShadingMode;
-
-// @public
-export type FieldShadingMode = 'never' | 'when-selected' | 'always';
-
-// @public
 export interface OverlayRect {
     readonly className?: string;
     // (undocumented)
@@ -43,14 +37,12 @@ export interface PaintOptions {
     // (undocumented)
     readonly drawingStrings?: DrawingPaintStrings;
     readonly emptyTocPlaceholderIds?: ReadonlySet<string>;
-    readonly fieldShading?: FieldShadingMode;
     readonly fontAlias?: (family: string) => string | undefined;
     // (undocumented)
     readonly imageUrlPort?: PaintImageUrlPort;
     readonly materialize?: ReadonlySet<number>;
     readonly readOnlyParagraphIds?: ReadonlySet<string>;
     readonly scale?: number;
-    readonly shadeFormFields?: boolean;
 }
 
 // @public
