+++ docs/api/docx-editor-core/store.api.md
@@ -974,9 +974,6 @@ export const DEFAULT_MAX_EDITABLE_STORY_PARTS = 64;
 export const DEFAULT_OOXML_PACKAGE_LIMITS: Required<Pick<OoxmlPackageLimits, 'maxXmlParts' | 'maxRelationships'>>;
 
 // @public
-export const DEFAULT_VIEW_SETTINGS: DocumentViewSettings;
-
-// @public
 export const DEFAULT_ZIP_LIMITS: ZipLimits;
 
 // @public
@@ -1092,11 +1089,6 @@ export interface DocumentTrackingSettings {
     readonly doNotTrackMoves: boolean;
     readonly restrictedToTrackedChanges: boolean;
     readonly trackRevisions: boolean;
-}
-
-// @public
-export interface DocumentViewSettings {
-    readonly doNotShadeFormData: boolean;
 }
 
 // @public
@@ -1370,9 +1362,6 @@ export function hasCommentPart(pkg: OoxmlPackage, storyPartName: string): boolea
 
 // @public
 export function hashAuthored(authoredState: unknown): string;
-
-// @public
-export function hasLegacyFormFieldData(node: OoxmlNode): boolean;
 
 // @public
 export function hasNode(part: OoxmlPart, nodeId: string): boolean;
@@ -3018,9 +3007,6 @@ export function readOoxmlPart(xml: string, metadata: OoxmlPartMetadata, limits?:
 
 // @public
 export function readTrackingSettings(settingsRoot: OoxmlNode | null | undefined): DocumentTrackingSettings;
-
-// @public
-export function readViewSettings(settingsRoot: OoxmlNode | null | undefined): DocumentViewSettings;
 
 // @public
 export function readXml(xml: string, limits?: XmlLimits): XmlResult;
